Updated Some Policy thing
Some weird google  Bullshit

So, our privacy policy URL is ready. What is next?

We have to place a privacy policy link within the app. You can create the same as a text link anywhere in the app menu. It should be opened in the external browser on click.
Then we have to place the same link in the Play Store. Go to Play Store console -> Store presence -> Store listing. Go to the last field named ‘Privacy Policy’ and update the link there.
Submit app update and Store listing update together.
Your app will be back online within 2 hours!
Hope it was helpful. Please discuss the issues, if any in comments so that I can try to help you further. Thanks for reading my post.